Improve compression performance
The performance is good for low depth frames or even depths because the Huffman tree constructed is small, when the depth is increased, there tend to be more variations with the depth, resulting in a larger Huffman tree that will take longer to process.
Filtering can be applied to the depth to remove these outlier depth points and reduce the size of the Huffman tree and improve performance.